TRAC#7506 Ensures registration requests are unique
