+++ app/comic/settings.py
@@ -13,7 +13,6 @@ from django.core.exceptions import ImproperlyConfigured
 from matplotlib.dates import hours
 
 DEBUG = True
-TEMPLATE_DEBUG = DEBUG
 
 ADMINS = (
     # ('Your Name', 'your_email@example.com'),
